Moved command enum into a separate module
