+++ hedgehog-tui/src/screen.rs
@@ -1,14 +1,15 @@
 use crate::cmdcontext::CommandContext;
 use crate::cmdreader::CommandReader;
+use crate::command::{Command, CommandConfirmation, FocusedPane, LinkType, Predicate};
 use crate::events::key;
 use crate::history::CommandsHistory;
-use crate::keymap::{Key, KeyMapping};
+use crate::keymap::KeyMapping;
 use crate::logger::{log_set_level, LogEntry, LogHistory, Severity};
 use crate::mouse::{MouseEventKind, MouseHitResult, MouseState, WidgetPositions};
-use crate::options::{Options, OptionsUpdate};
+use crate::options::Options;
 use crate::scrolling::pagination::{DataProvider, PaginatedData};
 use crate::scrolling::{selection, DataView, ScrollAction, ScrollableList};
-use crate::theming::{Theme, ThemeCommand};
+use crate::theming::Theme;
 use crate::widgets::animation::AnimationController;
 use crate::widgets::command::{CommandActionResult, CommandEditor, CommandState};
 use crate::widgets::confirmation::ConfirmationView;
@@ -24,7 +25,7 @@ use actix::prelude::*;
 use crossterm::event::{self, Event};
 use crossterm::QueueableCommand;
 use hedgehog_library::model::{
-    Episode, EpisodeId, EpisodePlaybackData, EpisodeStatus, EpisodeSummary, EpisodeSummaryStatus,
+    Episode, EpisodeId, EpisodePlaybackData, EpisodeSummary, EpisodeSummaryStatus,
     EpisodesListMetadata, Feed, FeedId, FeedSummary, FeedView, GroupId, GroupSummary, Identifiable,
 };
 use hedgehog_library::search::{self, SearchClient, SearchResult};
@@ -35,7 +36,6 @@ use hedgehog_library::{
     FeedUpdateRequest, FeedUpdateResult, Library, NewFeedMetadata, UpdateQuery,
 };
 use hedgehog_player::state::PlaybackState;
-use hedgehog_player::volume::VolumeCommand;
 use hedgehog_player::{
     InitialPlaybackState, PlaybackCommand, PlaybackMetadata, Player, PlayerNotification,
     SeekDirection, SeekOffset,
@@ -44,20 +44,8 @@ use std::collections::HashSet;
 use std::io::{stdout, Write};
 use std::iter::once;
 use std::ops::Range;
-use std::path::PathBuf;
 use std::time::Duration;
 use tui::backend::CrosstermBackend;
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, cmdparse::Parsable)]
-pub(crate) enum FocusedPane {
-    #[cmd(rename = "feeds")]
-    FeedsList,
-    #[cmd(rename = "episodes")]
-    EpisodesList,
-    Search,
-    #[cmd(rename = "log")]
-    ErrorsLog,
-}
 
 pub(crate) enum SearchState {
     Loading,
@@ -94,120 +82,6 @@ impl LibraryViewModel {
         self.episodes.set_window_size(window_size);
         self.feeds.set_window_size(window_size);
     }
-}
-
-#[derive(Debug, Clone, PartialEq, cmdparse::Parsable)]
-#[cmd(ctx = "CommandContext<'_>")]
-pub(crate) enum Command {
-    #[cmd(rename = "line")]
-    Cursor(ScrollAction),
-    Map(Key, #[cmd(attr(state))] Option<FocusedPane>, Box<Command>),
-    Unmap(Key, #[cmd(attr(state))] Option<FocusedPane>),
-    Theme(ThemeCommand),
-    Exec(PathBuf),
-    Confirm(Box<CommandConfirmation>),
-    #[cmd(transparent)]
-    Volume(VolumeCommand),
-    PlayCurrent,
-    #[cmd(transparent)]
-    Playback(PlaybackCommand),
-    Finish,
-    #[cmd(alias = "enable", alias = "disable")]
-    SetFeedEnabled(
-        #[cmd(
-            alias_value(alias = "enable", value = "true"),
-            alias_value(alias = "disable", value = "false")
-        )]
-        bool,
-    ),
-    #[cmd(alias = "q")]
-    Quit,
-    #[cmd(rename = "focus", alias = "log")]
-    SetFocus(#[cmd(alias_value(alias = "log", value = "FocusedPane::ErrorsLog"))] FocusedPane),
-    #[cmd(rename = "set")]
-    SetOption(OptionsUpdate),
-    #[cmd(rename = "add")]
-    AddFeed(String),
-    AddGroup(#[cmd(parser = "crate::cmdcontext::GroupNameParser")] String),
-    SetGroup(#[cmd(parser = "crate::cmdcontext::GroupNameParser")] String),
-    PlaceGroup(usize),
-    #[cmd(alias = "delete-feed")]
-    Delete,
-    Reverse,
-    Rename(#[cmd(parser = "hedgehog_library::search::SearchQueryParser")] String),
-    #[cmd(alias = "u")]
-    Update {
-        #[cmd(attr(this = "true"))]
-        current_only: bool,
-    },
-    AddArchive(String),
-    Mark {
-        status: EpisodeStatus,
-        #[cmd(attr(all = "true"))]
-        update_all: bool,
-        #[cmd(attr(if))]
-        condition: Option<EpisodeSummaryStatus>,
-    },
-    #[cmd(ignore, alias = "hide", alias = "unhide")]
-    SetEpisodeHidden(
-        #[cmd(
-            alias_value(alias = "hide", value = "true"),
-            alias_value(alias = "unhide", value = "false")
-        )]
-        bool,
-    ),
-    #[cmd(alias = "s")]
-    Search(#[cmd(parser = "hedgehog_library::search::SearchQueryParser")] String),
-    SearchAdd,
-    OpenLink(LinkType),
-
-    RepeatCommand,
-    Refresh,
-
-    Chain(Vec<Command>),
-    #[cmd(rename = "if")]
-    Conditional {
-        predicate: Predicate,
-        command: Box<Command>,
-        #[cmd(attr(else))]
-        otherwise: Option<Box<Command>>,
-    },
-    #[cmd(rename = "msg")]
-    WriteMessage {
-        message: String,
-        #[cmd(
-            default = "Severity::Information",
-            attr(
-                info = "Severity::Information",
-                warn = "Severity::Warning",
-                error = "Severity::Error",
-            )
-        )]
-        severity: Severity,
-    },
-}
-
-#[derive(Debug, Clone, cmdparse::Parsable, PartialEq)]
-pub(crate) enum Predicate {
-    Not(Box<Predicate>),
-    Either(Vec<Predicate>),
-    Both(Vec<Predicate>),
-    Focused(FocusedPane),
-}
-
-#[derive(Debug, Clone, PartialEq, cmdparse::Parsable)]
-pub(crate) enum LinkType {
-    Feed,
-    Episode,
-}
-
-#[derive(Debug, Clone, PartialEq, cmdparse::Parsable)]
-#[cmd(ctx = "CommandContext<'_>")]
-pub(crate) struct CommandConfirmation {
-    pub(crate) prompt: String,
-    pub(crate) action: Command,
-    #[cmd(attr(default))]
-    pub(crate) default: bool,
 }
 
 #[derive(Message)]
