+++ hedgehog-tui/src/widgets/library.rs
@@ -3,9 +3,10 @@ use super::empty::EmptyView;
 use super::episode_row::{EpisodesListRowRenderer, EpisodesListSizing};
 use super::feed_row::FeedsListRowRenderer;
 use super::list::List;
+use crate::command::FocusedPane;
 use crate::mouse::WidgetPositions;
 use crate::options::Options;
-use crate::screen::{FocusedPane, LibraryViewModel};
+use crate::screen::LibraryViewModel;
 use crate::scrolling::DataView;
 use crate::theming::{self, Theme};
 use hedgehog_library::model::{FeedStatus, FeedView};
